disp: adapt tests (mostly d2gr/d2p4 --> d2) and migrate dftd3 interface files
