+++ psi4/driver/qcdb/util/regex.py
