+++ tests/dftd3/grad/input.dat
@@ -23,7 +23,9 @@ set {
     d_convergence 8
 }
 
-grad = gradient('bp86-d2gr')
-compare_matrices(ref, grad, 7, "Outsourced dft gradients called by name")  #TEST
+grad = gradient('bp86-d2', engine='libdisp')
+compare_matrices(ref, grad, 7, "Outsourced dft gradients called by name: libdisp")  #TEST
 clean()
 
+grad = gradient('bp86-d2', engine='dftd3')
+compare_matrices(ref, grad, 7, "Outsourced dft gradients called by name: dftd3")  #TEST
